[REL] Migration purchase_multic_fix to 11.0
* [FIX] Update Readme & manifest
* [FIX] separe in diffent files the diferrent clases in the file .py
* [FIX] remove the executable in __init__.py files
