+++ purchase_multic_fix/models/__init__.py
@@ -2,4 +2,6 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import models
+from . import purchase_order
+from . import purchase_order_line
+from . import account_invoice
