+++ purchase_multic_fix/views/purchase_order_views.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="purchase_order_form" model="ir.ui.view">
+        <field name="name">purchase.order.form</field>
+        <field name="model">purchase.order</field>
+        <field name="inherit_id" ref="purchase.purchase_order_form"/>
+        <field name="arch" type="xml">
+            <field name="order_line" position="attributes">
+                <attribute name="context">{'company_id':company_id}</attribute>
+            </field>
+            <field name="taxes_id" position="attributes">
+                <!-- add company domain -->
+                <attribute name="domain">[('type_tax_use','=','purchase'),('company_id','=',parent.company_id)]</attribute>
+            </field>
+        </field>
+    </record>
+
+</odoo>
